requirements.txt
# python version 3.11

ortools
folium>=0.14.0
geopy>=2.3.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0
googlemaps>=4.10.0
python-dotenv>=1.0.0
